requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
assemblyai==0.20.0
google-generativeai==0.3.0
pydantic==2.5.0
typing-extensions==4.8.0
